refactor(analytics): avoid duplication in defining AnalyticsEventType
