+++ src/analytics/Events.tsx
@@ -347,7 +347,6 @@ export enum FeeEvents {
   fetch_tobin_tax_failed = 'fetch_tobin_tax_failed',
 }
 
-// Generic transaction logging to grab tx hashes
 export enum TransactionEvents {
   transaction_start = 'transaction_start',
   transaction_gas_estimated = 'transaction_gas_estimated',
@@ -639,37 +638,3 @@ export enum DappShortcutsEvents {
   dapp_shortcuts_reward_tx_accepted = 'dapp_shortcuts_reward_tx_accepted', // When the user confirms the transaction via the bottom sheet
   dapp_shortcuts_reward_tx_rejected = 'dapp_shortcuts_reward_tx_rejected', // When the user rejects the transaction via the bottom sheet
 }
-
-export type AnalyticsEventType =
-  | AppEvents
-  | HomeEvents
-  | SettingsEvents
-  | KeylessBackupEvents
-  | OnboardingEvents
-  | VerificationEvents
-  | IdentityEvents
-  | AuthenticationEvents
-  | InviteEvents
-  | EscrowEvents
-  | FiatExchangeEvents
-  | SendEvents
-  | RequestEvents
-  | FeeEvents
-  | TransactionEvents
-  | CeloExchangeEvents
-  | PerformanceEvents
-  | NavigationEvents
-  | RewardsEvents
-  | WalletConnectEvents
-  | DappKitEvents
-  | CICOEvents
-  | DappExplorerEvents
-  | WebViewEvents
-  | CoinbasePayEvents
-  | SwapEvents
-  | CeloNewsEvents
-  | TokenBottomSheetEvents
-  | AssetsEvents
-  | NftEvents
-  | BuilderHooksEvents
-  | DappShortcutsEvents
